modified lateral distance plot and modfied obsterm
added a new lateral distance plotter which is now much faster than the
previous we had but still there are a few small adjustments which need
to be made.
also modified the grouped obsterm so that when calling self.info we
correctly get the infos of all the obsterms which the groupedobsterm
holds. before we just thown away the infos