first fixes to the vec_env on linux
Changed the the lauch_game function and also the __get_tmnf_process_id
function for the linux game manager. The previous strategy of the
get_tmnf_process_id did not work since on linux the game does not launch
as a child process of TMLoader, so we swapped it for the old one. The
problem with the new impl. in the context of the vectorized approach is
that its not really Thread safe. When using multiple game processes they
will all call the function in parralell leading to everyone getting the
same pid. To circumentvent that we added a Lock such that only one
instance can get their pid.